% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{approxErrorPropagation}
\alias{approxErrorPropagation}
\title{Approximative error propagation rule}
\usage{
approxErrorPropagation(dVal, var)
}
\arguments{
\item{dVal}{Values of the partial derived functions with arguments inserted}

\item{var}{Variances for the given values}
}
\value{
The error propagation
}
\description{
Calculates the non-linear approximative error propagation rule (Method 4.3)
}
