% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{beta0hat}
\alias{beta0hat}
\title{Simple linear regression estimator beta0}
\usage{
beta0hat(x, y)
}
\arguments{
\item{x}{Vector of data}

\item{y}{Vector of data}
}
\value{
The estimator beta0 hat
}
\description{
Theorem 5.4
}
