% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability.R
\name{binomMean}
\alias{binomMean}
\title{Calculates the mean of a binomial distributed random variable}
\usage{
binomMean(n, p)
}
\arguments{
\item{n}{number of independent draws}

\item{p}{probability of success}
}
\description{
Calculates the mean of a binomial distributed random variable
}
