% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability.R
\name{binomVar}
\alias{binomVar}
\title{Calculates the variance of a binomial distributed random variable}
\usage{
binomVar(n, p)
}
\arguments{
\item{n}{number of independent draws}

\item{p}{probability of success}
}
\description{
Calculates the variance of a binomial distributed random variable
}
