% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{bonferroni}
\alias{bonferroni}
\title{The bonferroni corrected alpha}
\usage{
bonferroni(numOfGroups, alpha = 0.05)
}
\arguments{
\item{numOfGroups}{number of groups}

\item{alpha}{Significance level (0.05 by default)}
}
\value{
The bonferroni corrected significance level
}
\description{
The bonferroni corrected alpha
}
