% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisOfTwoSamples.R
\name{confInterval2Raw}
\alias{confInterval2Raw}
\title{Two sample confidence interval for mean1 - mean2}
\usage{
confInterval2Raw(m, v, n, alpha = 0.05)
}
\arguments{
\item{m}{Vector of 2 means}

\item{v}{Vector of 2 variances}

\item{n}{Vector of 2 sample sizes}

\item{alpha}{Significance level (0.05 by default)}
}
\value{
Confidence interval for the difference in mean of x and y with significance level alpha
}
\description{
Calculates the confidence interval for two samples using raw data
}
