% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSS.R
\name{confIntLine}
\alias{confIntLine}
\title{Confidence interval for line}
\usage{
confIntLine(fit, xnew, alpha = 0.05)
}
\arguments{
\item{fit}{linear model}

\item{xnew}{new data}

\item{alpha}{Significance level (0.05 by default)}
}
\value{
Confidence interval for new data
}
\description{
The line beta0 + beta1*xnew  \cr
Method 5.18
}
