% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{cv}
\alias{cv}
\title{Calculates the coefficient of variation (CV)}
\usage{
cv(val, digits = 2)
}
\arguments{
\item{val}{The value to calculate}

\item{digits}{The amount of digits to print}
}
\value{
A percentage representing the CV of val
}
\description{
The CV is the standard deviation seen relative to the mean of a given value
}
