% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{df2}
\alias{df2}
\title{Method 3.51
Two sample degrees of freedom}
\usage{
df2(x, y)
}
\arguments{
\item{x}{Vector of data}

\item{y}{Vector of data}
}
\value{
The v degrees of freedom for two sample
}
\description{
Method 3.51
Two sample degrees of freedom
}
