% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisOfTwoSamples.R
\name{dof2Raw}
\alias{dof2Raw}
\title{Raw dof of two samples}
\usage{
dof2Raw(v, n)
}
\arguments{
\item{v}{Vector of 2 variances}

\item{n}{Vector of 2 sample sizes}
}
\value{
The degrees of freedom
}
\description{
Calculates the degrees of freedom for two samples using raw data
}
