% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{Fobs}
\alias{Fobs}
\title{Calculate the test statistic F_obs}
\usage{
Fobs(x, treatments)
}
\arguments{
\item{x}{Vector of all data from all groups}

\item{treatments}{Vector with pointers for which group each entry in x belong to.}
}
\value{
The test statistic F_obs
}
\description{
Theorem 8.6 \cr
One-way ANOVA
}
