% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{FTr}
\alias{FTr}
\title{Theorem 8.22
Two-way ANOVA
Treatment test statistic Ftr}
\usage{
FTr(x, treatments, blocks)
}
\arguments{
\item{x}{Vector of all data from all groups}

\item{treatments}{Vector with pointers for which group each entry in x belong to.}

\item{blocks}{Vector with pointers for which block each entry in x belong to}
}
\value{
The test statistic F for treatments
}
\description{
Theorem 8.22
Two-way ANOVA
Treatment test statistic Ftr
}
