% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability.R
\name{hyperMean}
\alias{hyperMean}
\title{Calculates the mean of a hypergeometric distributed variable}
\usage{
hyperMean(a, N, n)
}
\arguments{
\item{a}{number of successes}

\item{N}{population size}

\item{n}{amount of draws}
}
\description{
Calculates the mean of a hypergeometric distributed variable
}
