% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability.R
\name{hyperVar}
\alias{hyperVar}
\title{Calculates the variance of a hypergeometric distributed variable}
\usage{
hyperVar(a, N, n)
}
\arguments{
\item{a}{number of successes}

\item{N}{population size}

\item{n}{amount of draws}
}
\description{
Calculates the variance of a hypergeometric distributed variable
}
