% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{linearAnal}
\alias{linearAnal}
\title{Simple linear regression analysis}
\usage{
linearAnal(y, x)
}
\arguments{
\item{y}{Vector of data}

\item{x}{Vector of data}
}
\value{
Summary of the linear model given by y ~ x
}
\description{
Linear model based on y ~ x
}
