% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPackage.R
\name{lsp}
\alias{lsp}
\title{Searching for functions in a package}
\usage{
lsp(package, pattern)
}
\arguments{
\item{package}{name of package to search}

\item{pattern}{find all functions with pattern in the name}
}
\value{
all functions in package
}
\description{
Searching for functions in a package
}
