% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginOfError.R
\name{ME}
\alias{ME}
\title{Margin of error}
\usage{
ME(x, alpha = 0.05)
}
\arguments{
\item{x}{Vector of data}

\item{alpha}{Significance level (0.05 by default)}
}
\value{
Half the width of the confidence interval
}
\description{
The margin of error, also called half the width of the confidence interval
}
