% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{MSE}
\alias{MSE}
\title{One- and Two-way ANOVA}
\usage{
MSE(x, treatments, blocks = NULL)
}
\arguments{
\item{x}{Vector of all data from all groups}

\item{treatments}{Vector with pointers for which group each entry in x belong to.}

\item{blocks}{Vector with pointers for which block each entry in x belong to (Null if One-way ANOVA)}
}
\value{
The mean sum of squares
}
\description{
One- and Two-way ANOVA
}
