% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{MSTr}
\alias{MSTr}
\title{One-way ANOVA}
\usage{
MSTr(x, treatments)
}
\arguments{
\item{x}{Vector of all data from all groups}

\item{treatments}{Vector with pointers for which group each entry in x belong to.}
}
\value{
The mean treatment sum of squares
}
\description{
One-way ANOVA
}
