% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSS.R
\name{paraConfInt}
\alias{paraConfInt}
\title{Parameter estimator confidence interval}
\usage{
paraConfInt(beta, sigmabeta, n, alpha = 0.05)
}
\arguments{
\item{beta}{Estimator parameter beta}

\item{sigmabeta}{Variance of estimator beta}

\item{n}{length of vector}

\item{alpha}{Significance level (0.05 by default)}
}
\value{
Confidence interval for estimator parameter
}
\description{
Method 5.15
}
