% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proportions.R
\name{propConfInt}
\alias{propConfInt}
\title{Proportion confidence interval}
\usage{
propConfInt(x, n, alpha = 0.05)
}
\arguments{
\item{x}{Number of observations in a category}

\item{n}{Number of observations in total}

\item{alpha}{Significance level (0.05 by default)}
}
\value{
The proportion confidence interval
}
\description{
Method 7.3
}
