% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{pval2}
\alias{pval2}
\title{Two sample p-value
or just do t.test(x,y)}
\usage{
pval2(x, y, mu0)
}
\arguments{
\item{x}{Vector of data}

\item{y}{Vector of data}

\item{mu0}{The null hypothesis mean value}
}
\value{
The t-test p-value
}
\description{
Two sample p-value
or just do t.test(x,y)
}
