% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisOfTwoSamples.R
\name{pval2Raw}
\alias{pval2Raw}
\title{Raw two sample p-value}
\usage{
pval2Raw(m, v, n)
}
\arguments{
\item{m}{Vector of 2 means}

\item{v}{Vector of 2 variances}

\item{n}{Vector of 2 sample sizes}
}
\value{
The t-test p-value
}
\description{
Calculates the p-val for two samples using raw data
}
