% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{pvalANOVA}
\alias{pvalANOVA}
\title{Calculate the p-value for One-way ANOVA}
\usage{
pvalANOVA(x, treatments)
}
\arguments{
\item{x}{Vector of all data from all groups}

\item{treatments}{Vector with pointers for which group each entry in x belong to.}
}
\value{
The test p-value
}
\description{
Calculate the p-value for One-way ANOVA
}
