% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSS.R
\name{pvalBeta}
\alias{pvalBeta}
\title{p-value for simple linear regression}
\usage{
pvalBeta(tobs, n)
}
\arguments{
\item{tobs}{The test statistic tobs for estimator beta}

\item{n}{length of vector}
}
\value{
p-value
}
\description{
Method 5.14 \cr
or just use confint(fit, level=0.95)
}
