% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{pvalPostHoc}
\alias{pvalPostHoc}
\title{Post hoc pairwise hypothesis test p-value for One- and Two-way ANOVA}
\usage{
pvalPostHoc(
  x,
  treatments,
  groupIndex1,
  groupIndex2,
  blocks = NULL,
  alpha = 0.05
)
}
\arguments{
\item{x}{Vector of data}

\item{treatments}{Vector with pointers for which group each entry in x belong to.}

\item{groupIndex1}{Index of group to compare}

\item{groupIndex2}{Index of group to compare}

\item{alpha}{Significance level (0.05 by default)}
}
\value{
The p-value for post hoc pairwise hypothesis test
}
\description{
Method 8.10
}
