% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{rawdf2}
\alias{rawdf2}
\title{Two sample degrees of freedom}
\usage{
rawdf2(v, n)
}
\arguments{
\item{v}{Vector of variance}

\item{n}{Vector of length}
}
\value{
The v degrees of freedom for two sample
}
\description{
Method 3.51 \cr
Calculated from given summary statistics
}
