% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginOfError.R
\name{rawME}
\alias{rawME}
\title{Margin of error}
\usage{
rawME(s, n, alpha = 0.05)
}
\arguments{
\item{s}{Standard deviation of a vector}

\item{n}{Sample size}

\item{alpha}{Significance level (0.05 by default)}
}
\value{
Half the width of the confidence interval
}
\description{
The margin of error, also called half the width of the confidence interval \cr
Calculated from raw summary statistics
}
