% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceInterval.R
\name{rawMeanInterval}
\alias{rawMeanInterval}
\title{Calculates the mean confidence interval without a vector}
\usage{
rawMeanInterval(mu, sd, len, alpha = 0.05)
}
\arguments{
\item{mu}{Given mean}

\item{sd}{Standard deviation}

\item{len}{Length of the vector}

\item{alpha}{Significance level}
}
\value{
The mean confidence interval
}
\description{
Calculates the mean confidence interval without a vector
}
