% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proportions.R
\name{rawPropConfInt}
\alias{rawPropConfInt}
\title{Proportion confidence interval}
\usage{
rawPropConfInt(p, alpha = 0.05)
}
\arguments{
\item{p}{Proportion given by x/n}

\item{alpha}{Significance level (0.05 by default)}
}
\value{
The proportion confidence interval
}
\description{
Method 7.3
}
