% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceInterval.R
\name{rawSdInterval}
\alias{rawSdInterval}
\title{Calculates the standard deviation confidence interval without a vector}
\usage{
rawSdInterval(sd, len, alpha = 0.05)
}
\arguments{
\item{sd}{Given standard deviation}

\item{len}{Length of the vector}

\item{alpha}{Significance level}
}
\value{
The standard deviation confidence interval
}
\description{
Calculates the standard deviation confidence interval without a vector
}
