% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceInterval.R
\name{rawVarInterval}
\alias{rawVarInterval}
\title{Calculates the variance confidence interval without a vector}
\usage{
rawVarInterval(var, len, alpha = 0.05)
}
\arguments{
\item{var}{Given variance}

\item{len}{Length of the vector}

\item{alpha}{Significance level}
}
\value{
The variance confidence interval
}
\description{
Calculates the variance confidence interval without a vector
}
