% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSS.R
\name{RSS}
\alias{RSS}
\title{Residual sum of squares}
\usage{
RSS(y, x)
}
\arguments{
\item{y}{Vector of data}

\item{x}{Vector of data}
}
\value{
Sum of the residuals of the linear model lm(y~x) squared
}
\description{
Input makes a simple linear model lm(y~x)
}
