% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginOfError.R
\name{sampleSizeME}
\alias{sampleSizeME}
\title{Confidence interval sample size}
\usage{
sampleSizeME(ME, s, alpha = 0.05)
}
\arguments{
\item{ME}{Margin of error}

\item{s}{Standard deviation of data}

\item{alpha}{Significance level (0.05 by default)}
}
\value{
The minimum sample size needed for a given margin of error
}
\description{
Calculate the sample size needed for a given margin of error \cr
Method 3.63.
}
