% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{sdInterval}
\alias{sdInterval}
\title{Method 3.19
Standard deviation interval}
\usage{
sdInterval(x, alpha = 0.05)
}
\arguments{
\item{x}{Vector of data}

\item{alpha}{Significance level (0.05 by default)}
}
\value{
The confidence interval for the standard deviation
}
\description{
Method 3.19
Standard deviation interval
}
