% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceInterval.R
\name{SEM}
\alias{SEM}
\title{Calculates the sample error of the mean of a set}
\usage{
SEM(x)
}
\arguments{
\item{x}{The set to calculate}
}
\value{
The standard error of the mean
}
\description{
Definition 3.7
}
