% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSS.R
\name{sigmaBeta0}
\alias{sigmaBeta0}
\title{Variation of estimator beta0}
\usage{
sigmaBeta0(x, RSS)
}
\arguments{
\item{x}{Vector of data}

\item{RSS}{Residual sum of squares}
}
\value{
The variation of estimator beta0
}
\description{
Theorem 5.8
}
