% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSS.R
\name{sigmaBeta1}
\alias{sigmaBeta1}
\title{Variation of estimator beta1}
\usage{
sigmaBeta1(x, RSS)
}
\arguments{
\item{x}{Vector of data}

\item{RSS}{Residual sum of squares}
}
\value{
The variation of estimator beta1
}
\description{
Theorem 5.8
}
