% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSS.R
\name{sigmaSquare}
\alias{sigmaSquare}
\title{Variance of estimators}
\usage{
sigmaSquare(RSS, n)
}
\arguments{
\item{RSS}{Residual sum of squares}

\item{n}{length of vector}
}
\value{
The variance of estimators
}
\description{
Simple linear regression variance of estimators beta0 and beta1
}
