% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{SSE}
\alias{SSE}
\title{Residual sum of squares}
\usage{
SSE(x, treatments, blocks = NULL)
}
\arguments{
\item{x}{Vector of all data from all groups}

\item{treatments}{Vector with pointers for which group each entry in x belong to.}

\item{blocks}{Vector with pointers for which block each entry in x belong to (Null if One-way ANOVA)}
}
\value{
The sum of squared errors
}
\description{
Theorem 8.2 and 8.20 \cr
One- and Two-way ANOVA
}
