% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{SST}
\alias{SST}
\title{Theorem 8.2
One-way ANOVA
Can also be calculated as SSTr + SSE}
\usage{
SST(x)
}
\arguments{
\item{x}{Vector of data}
}
\value{
The total sum of squares
}
\description{
Theorem 8.2
One-way ANOVA
Can also be calculated as SSTr + SSE
}
