% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{SSTr}
\alias{SSTr}
\title{Treatment sum of squares}
\usage{
SSTr(x, treatments)
}
\arguments{
\item{x}{Vector of all data from all groups}

\item{treatments}{Vector with pointers for which group each entry in x belong to.}
}
\value{
The treatment sum of squares
}
\description{
Treatment sum of squares
}
