% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{Sxx}
\alias{Sxx}
\title{The sample correct sum of squares}
\usage{
Sxx(x)
}
\arguments{
\item{x}{Vector of data}
}
\value{
The sum of squares of the difference between x and its mean
}
\description{
Theorem 5.4
}
