% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{tobs2}
\alias{tobs2}
\title{Method 3.51
Two sample t_obs value}
\usage{
tobs2(x, y, mu0 = 0)
}
\arguments{
\item{x}{Vector of data}

\item{y}{Vector of data}

\item{mu0}{The null hypothesis mean value}
}
\value{
The test statistic t_obs for two sample t-test
}
\description{
Method 3.51
Two sample t_obs value
}
