% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisOfTwoSamples.R
\name{tobs2Raw}
\alias{tobs2Raw}
\title{Raw tobs of two samples}
\usage{
tobs2Raw(m, v, n)
}
\arguments{
\item{m}{Vector of 2 means}

\item{v}{Vector of 2 variances}

\item{n}{Vector of 2 sample sizes}
}
\value{
The t-obs value
}
\description{
Calculates the tobs for two samples using raw data
}
