% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSS.R
\name{tobsBeta}
\alias{tobsBeta}
\title{Test statistic tobs for estimator beta}
\usage{
tobsBeta(betaI, beta0I, sigmabetaI)
}
\arguments{
\item{betaI}{Estimator betaI}

\item{beta0I}{Null hypothesis betaI = beta0I}

\item{sigmabetaI}{Variation of betaI}
}
\value{
The test statistic tobs for beta_i
}
\description{
Null hypothesis test (betaI = beta0I) \cr
Theorem 5.12
}
