% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{varInterval}
\alias{varInterval}
\title{Variance interval}
\usage{
varInterval(x, alpha = 0.05)
}
\arguments{
\item{x}{Vector of data}

\item{alpha}{Significance level (0.05 by default)}
}
\value{
The confidence interval for the variation deviation
}
\description{
Method 3.19
}
