% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{wally}
\alias{wally}
\title{Wally plot is a guessing game where your data is compared with 8 normally distributed data.
Pick the data that deviates most from the line, if it is not yours then your data is assumed to be normally distributed
You cannot conclude anything from wallyplot, it is merely a tool}
\usage{
wally(x, range1, range2)
}
\arguments{
\item{x}{Vector of data to plot}

\item{range1}{y-axis plot range lower value}

\item{range2}{y-axis plot range upper value}
}
\value{
Wallyplot of the data compared with 8 normally distributed data
}
\description{
Wally plot is a guessing game where your data is compared with 8 normally distributed data.
Pick the data that deviates most from the line, if it is not yours then your data is assumed to be normally distributed
You cannot conclude anything from wallyplot, it is merely a tool
}
