% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proportions.R
\name{zobs}
\alias{zobs}
\title{The test statistic Z_obs}
\usage{
zobs(x, p0, n)
}
\arguments{
\item{x}{Number of observations in a category}

\item{p0}{The null hypothesis}

\item{n}{Number of observations in total}
}
\value{
Z_obs test statistic
}
\description{
Method 7.11 \cr
Used for hypothesis test \cr
Use prop.test(x = x, n = n, p = p0, correct = FALSE, conf.level = 1-alpha) to get p-value
}
