% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proportions.R
\name{zobs2}
\alias{zobs2}
\title{The test statistic Z_obs for two sample proportions}
\usage{
zobs2(x, p, n)
}
\arguments{
\item{x}{Vector of two numbers of observations in a category}

\item{n}{Vector of two number of observations in total}

\item{p0}{Vector of two number of the null hypothesis}
}
\value{
Z_obs test statistic for two sample proportions hypothesis test
}
\description{
Method 7.18 \cr
Used for two sample proportions hypothesis test \cr
Use prop.test(x = c(x1, x2), n = c(n1,n2), p = c(p1,p2), correct = FALSE, conf.level = 1-alpha) to get p-value
}
