% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proportions.R
\name{zval}
\alias{zval}
\title{The Z-distribution value}
\usage{
zval(alpha)
}
\arguments{
\item{alpha}{Significance level (0.05 by default)}
}
\value{
Z-distribution value
}
\description{
The Z-distribution is a normal distribution with mean zero and standard deviation 1
}
