% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{beta1hat}
\alias{beta1hat}
\title{Theorem 5.4
Linear regression estimator beta1}
\usage{
beta1hat(x, y)
}
\arguments{
\item{x}{Vector of data}

\item{y}{Vector of data}
}
\value{
The estimator beta1 hat
}
\description{
Theorem 5.4
Linear regression estimator beta1
}
