% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{postHocConfInt}
\alias{postHocConfInt}
\title{Post Hoc pairwise confidence interval}
\usage{
postHocConfInt(
  x,
  treatments,
  groupIndex1,
  groupIndex2,
  blocks = NULL,
  alpha = 0.05
)
}
\arguments{
\item{x}{Vector of data}

\item{treatments}{Vector with pointers for which group each entry in x belong to.}

\item{groupIndex1}{Index of group to compare}

\item{groupIndex2}{Index of group to compare}

\item{blocks}{Vector with pointers for which block each entry in x belong to (NULL if One-way ANOVA)}

\item{alpha}{Significance level (0.05 by default)}
}
\value{
Confidence interval for Post Hoc pairwise comparison
}
\description{
Post Hoc pairwise confidence interval
}
