% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{pval1}
\alias{pval1}
\title{The one-sample t-test p-value}
\usage{
pval1(x, mu = 0)
}
\arguments{
\item{x}{Vector of data}

\item{mu}{mean for the null hypothesis}
}
\value{
The t-test p-value
}
\description{
Method 3.23 \cr
or just do t.test(x, conf.level=0.95) :)
}
