% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KasperFuncs.R
\name{tobs1}
\alias{tobs1}
\title{Method 3.23
Calculate the one-sample t-test statistic}
\usage{
tobs1(x, mu = 0)
}
\arguments{
\item{x}{Vector of data}

\item{mu}{mean for the null hypothesis}
}
\value{
The test statistic t_obs
}
\description{
Method 3.23
Calculate the one-sample t-test statistic
}
